Demote items in map from float to int

+++ src/core.py
@@ -6,4 +6,4 @@ def square_function(xmin, xmax) -> map:
     """
     Return the range of f:[xmin,xmax] ⟶ ℝ, x ↦ x²
     """
-    return map(lambda x: math.pow(2, x), range(xmin, xmax+1))
+    return map(lambda x: int(math.pow(2, x)), range(xmin, xmax+1))
